Module google_workflows1::api

Structs

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

ListLocationsResponse
The response message for Locations.ListLocations.

ListOperationsResponse
The response message for Operations.ListOperations.

ListWorkflowsResponse
Response for the ListWorkflows method.

Location
A resource that represents Google Cloud Platform location.

Operation
This resource represents a long-running operation that is the result of a network API call.

ProjectLocationGetCall
Gets information about a location.

ProjectLocationListCall
Lists information about the supported locations for this service.

ProjectLocationOperationDeleteCall
Deletes a long-running operation. This method indicates that the client is no longer interested in the operation result. It does not cancel the operation. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED.

ProjectLocationOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

ProjectLocationOperationListCall
Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

ProjectLocationWorkflowCreateCall
Creates a new workflow. If a workflow with the specified name already exists in the specified project and location, the long running operation will return ALREADY_EXISTS error.

ProjectLocationWorkflowDeleteCall
Deletes a workflow with the specified name. This method also cancels and deletes all running executions of the workflow.

ProjectLocationWorkflowGetCall
Gets details of a single Workflow.

ProjectLocationWorkflowListCall
Lists Workflows in a given project and location. The default order is not specified.

ProjectLocationWorkflowPatchCall
Updates an existing workflow. Running this method has no impact on already running executions of the workflow. A new revision of the workflow may be created as a result of a successful update operation. In that case, such revision will be used in new workflow executions.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the Workflows hub.

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

Workflow
Workflow program to be executed by Workflows.

Workflows
Central instance to access all Workflows related resource activities

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.